fix(combat): migrate combat stat pipeline from u32 to u64 (#619)
* fix(combat): migrate combat stat pipeline from u32 to u64 (#575)

Enemy stats in endgame Loom zones exceed u32 range: Zone 50 base HP
(~3.28B) with depth offsets and the 5.0x zone boss multiplier
saturates at u32::MAX from ~Zone 44 onward, so every zone boss in
Z45-Z50 had identical clamped stats, flattening endgame progression.

Migrate the full combat stat pipeline to u64:
- Enemy (max_hp/current_hp/damage/defense), CombatState player HP
- DerivedStats (max_hp/physical_damage/magic_damage/defense)
- PrestigeCombatBonuses and CombatBonuses flat bonuses
- CombatEvent/TickEvent damage and heal payloads
- ZONE_ENEMY_STATS table and all enemy generators
- Player/enemy attack pipelines, regen, tick HP application,
  power rating, UI HP formatting, mkstate fixtures

Saved games remain compatible: JSON u32 values deserialize into u64
fields unchanged.

Adds regression tests asserting Zone 50 boss HP exceeds u32::MAX and
that Z43-Z50 zone boss HP strictly increases (both fail under u32
saturation).

Closes #575


* fix(tests): drop now-redundant u64 casts flagged by clippy

CI's clippy (1.96) flags `as u64` on enemy stat fields that are
already u64 after the migration (unnecessary_cast).


---------

+++ src/character/combat_bonuses.rs
@@ -7,13 +7,13 @@ use crate::core::constants::*;
 #[derive(Debug, Clone, Copy, Default)]
 pub struct PrestigeCombatBonuses {
     /// Flat damage added after Haven % multiplier, before crit
-    pub flat_damage: u32,
+    pub flat_damage: u64,
     /// Flat defense added to DEX-based defense
-    pub flat_defense: u32,
+    pub flat_defense: u64,
     /// Percentage points added to crit chance
     pub crit_chance: f64,
     /// Flat HP added to combat HP (NOT to DerivedStats.max_hp)
-    pub flat_hp: u32,
+    pub flat_hp: u64,
 }
 
 impl PrestigeCombatBonuses {
@@ -24,13 +24,13 @@ impl PrestigeCombatBonuses {
         Self {
             flat_damage: (PRESTIGE_FLAT_DAMAGE_FACTOR
                 * (rank as f64).powf(PRESTIGE_FLAT_DAMAGE_EXPONENT))
-            .floor() as u32,
+            .floor() as u64,
             flat_defense: (PRESTIGE_FLAT_DEFENSE_FACTOR
                 * (rank as f64).powf(PRESTIGE_FLAT_DEFENSE_EXPONENT))
-            .floor() as u32,
+            .floor() as u64,
             crit_chance: (rank as f64 * PRESTIGE_CRIT_PER_RANK).min(PRESTIGE_CRIT_CAP),
             flat_hp: (PRESTIGE_FLAT_HP_FACTOR * (rank as f64).powf(PRESTIGE_FLAT_HP_EXPONENT))
-                .floor() as u32,
+                .floor() as u64,
         }
     }
 }
